Optimize some components 2
